fix: align datatype with db

+++ models/wallet.py
@@ -7,4 +7,4 @@ class Wallet(Base):
     # only either `enabled_at` or `disabled_at` must have value. the other one must be null
     enabled_at = TimestampField(null=True)
     disabled_at = TimestampField(null=True)
-    balance: int = 0
+    balance: float = 0
